Tests/LibWeb/Text/expected/css/placeholder-custom-properties.txt

color: rgb(0, 0, 255)
opacity: 0.5
font-weight: 900
inherited color: rgb(0, 128, 0)
nested color: rgb(0, 0, 255)
